pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qr-token-stl-generator"
version = "0.1.0"
description = "Generate 3D printable STL files for QR code tokens from JSON export"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "qrcode>=7.4",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "numpy-stl>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
qr-token-stl = "app.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
